+++ packages/ui/src/common/hooks/useHttp/utils/request.ts
@@ -26,6 +26,7 @@ export const request = async (
       data,
       timeout,
       withCredentials: true,
+      params: request.params,
     };
 
     const response = await axios(config);