+++ services/workflows-service/src/common/utils/unified-api-client/unified-api-client.ts
@@ -15,6 +15,8 @@ import { Business, Customer } from '@prisma/client';
 import { TSchema } from '@sinclair/typebox';
 import { FEATURE_LIST, TCustomerWithFeatures } from '@/customer/types';
 import { TCustomerConfig } from '@/customer/schemas/zod-schemas';
+import { UpdateableAssessmentStatus } from '@ballerine/common';
+import { PageDto } from '@/common/dto';
 import { isType } from '@ballerine/common';
 import z from 'zod';
 
@@ -203,10 +205,54 @@ export class UnifiedApiClient {
     });
   }
 
+  public async getAssessmentsByType(
+    assessmentType: 'kyb_and_ownership',
+    projectId: string,
+    queryParams: {
+      page: number;
+      limit: number;
+    },
+  ) {
+    return await this.axiosInstance.get(`/assessments/${assessmentType}`, {
+      params: {
+        ...queryParams,
+        projectId,
+      },
+    });
+  }
+
+  public async getAssessmentById(id: string, projectId: string) {
+    return await this.axiosInstance.get(`/assessments/by-id/${id}?projectId=${projectId}`);
+  }
+
+  public async createAssessment(
+    assessmentType: 'kyb_and_ownership',
+    payload: {
+      registrationNumber: string;
+      companyName: string;
+      country: string;
+      projectId: string;
+      businessId?: string;
+    },
+  ) {
+    return await this.axiosInstance.post(`/assessments/${assessmentType}`, payload);
+  }
+
+  public async updateAssessmentStatus(
+    id: string,
+    status: UpdateableAssessmentStatus,
+    projectId: string,
+  ) {
+    return await this.axiosInstance.put(`/assessments/${id}/status`, {
+      status,
+      projectId,
+    });
+  }
+
   public async getLatestAssessmentsByWorkflowRuntimeDataId({
-    workflowRuntimeDataId,
-    projectId,
-  }: {
+                                                             workflowRuntimeDataId,
+                                                             projectId,
+                                                           }: {
     workflowRuntimeDataId: string;
     projectId: string;
   }) {