+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,107 @@
 # @ballerine/workflows-service
 
+## 0.7.138
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.107
+  - @ballerine/workflow-core@0.6.130
+  - @ballerine/workflow-node-sdk@0.6.130
+
+## 0.7.137
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.106
+  - @ballerine/workflow-core@0.6.129
+  - @ballerine/workflow-node-sdk@0.6.129
+
+## 0.7.136
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.105
+  - @ballerine/workflow-core@0.6.128
+  - @ballerine/workflow-node-sdk@0.6.128
+
+## 0.7.135
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.104
+  - @ballerine/workflow-core@0.6.127
+  - @ballerine/workflow-node-sdk@0.6.127
+
+## 0.7.134
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.103
+  - @ballerine/workflow-core@0.6.126
+  - @ballerine/workflow-node-sdk@0.6.126
+
+## 0.7.133
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.102
+  - @ballerine/workflow-core@0.6.125
+  - @ballerine/workflow-node-sdk@0.6.125
+
+## 0.7.132
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.101
+  - @ballerine/workflow-core@0.6.124
+  - @ballerine/workflow-node-sdk@0.6.124
+
+## 0.7.131
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.100
+  - @ballerine/workflow-core@0.6.123
+  - @ballerine/workflow-node-sdk@0.6.123
+
+## 0.7.130
+
+### Patch Changes
+
+- Bump
+
+## 0.7.129
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.122
+  - @ballerine/workflow-node-sdk@0.6.122
+  - @ballerine/common@0.9.99
+
+## 0.7.128
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.98
+  - @ballerine/workflow-core@0.6.121
+  - @ballerine/workflow-node-sdk@0.6.121
+
 ## 0.7.127
 
 ### Patch Changes